fix: 🐛 table issues
