+++ packages/plugins/preset-gfm/src/composed/schema.ts
@@ -1,12 +1,22 @@
 import type { MilkdownPlugin } from '@milkdown/ctx'
 import { strikethroughAttr, strikethroughSchema } from '../mark'
-import { extendListItemSchemaForTask, footnoteDefinitionSchema, footnoteReferenceSchema, tableCellSchema, tableHeaderSchema, tableRowSchema, tableSchema } from '../node'
+import {
+  extendListItemSchemaForTask,
+  footnoteDefinitionSchema,
+  footnoteReferenceSchema,
+  tableCellSchema,
+  tableHeaderRowSchema,
+  tableHeaderSchema,
+  tableRowSchema,
+  tableSchema,
+} from '../node'
 
 /// @internal
 export const schema: MilkdownPlugin[] = [
   extendListItemSchemaForTask,
 
   tableSchema,
+  tableHeaderRowSchema,
   tableRowSchema,
   tableHeaderSchema,
   tableCellSchema,
