+++ packages/plugins/preset-gfm/src/index.ts
@@ -6,4 +6,4 @@ export * from './plugin'
 export * from './composed'
 
 /// The GFM preset, includes all the plugins.
-export const gfm = [schema, inputRules, markInputRules, keymap, plugins, commands].flat()
+export const gfm = [schema, inputRules, markInputRules, keymap, commands, plugins].flat()
