+++ packages/plugins/preset-gfm/src/plugin/index.ts
@@ -1,5 +1,5 @@
-export * from './auto-insert-zero-space-plugin'
 export * from './auto-insert-span-plugin'
 export * from './column-resizing-plugin'
 export * from './table-editing-plugin'
 export * from './remark-gfm-plugin'
+export * from './keep-table-align-plugin'
