+++ packages/prose/package.json
@@ -180,7 +180,7 @@
     "prosemirror-model": "^1.19.3",
     "prosemirror-schema-list": "^1.3.0",
     "prosemirror-state": "^1.4.3",
-    "prosemirror-tables": "^1.3.4",
+    "prosemirror-tables": "^1.3.7",
     "prosemirror-transform": "^1.7.5",
     "prosemirror-view": "^1.31.7",
     "tslib": "^2.5.0"
